backend/requirements.txt
openai
promptlayer
python-dotenv
